update tests, fix integration text ini
