+++ tests/conftest.py
@@ -1,5 +1,7 @@
 """Shared test fixtures for the test suite."""
 
+from configparser import ConfigParser, ExtendedInterpolation
+
 import pytest
 
 from nsidc.metgen.models import CollectionMetadata
@@ -15,3 +17,23 @@ def simple_collection_metadata():
     return CollectionMetadata(
         short_name="ABCD", version="2", entry_title="Test Collection ABCD V002"
     )
+
+
+@pytest.fixture
+def cfg_parser():
+    cp = ConfigParser(interpolation=ExtendedInterpolation())
+    cp["Source"] = {"data_dir": "/data/example"}
+    cp["Collection"] = {"auth_id": "DATA-0001", "version": 42, "provider": "FOO"}
+    cp["Destination"] = {
+        "local_output_dir": "/output/here",
+        "ummg_dir": "ummg",
+        "kinesis_stream_name": "xyzzy-${environment}-stream",
+        "staging_bucket_name": "xyzzy-${environment}-bucket",
+        "write_cnm_file": False,
+    }
+    cp["Settings"] = {
+        "checksum_type": "SHA256",
+        "number": 1,
+        "log_dir": "/tmp",
+    }
+    return cp
