+++ tests/integration/configs/OLVIS1A_DUCk.ini
@@ -7,7 +7,7 @@ spatial_dir = ./data/spatial
 auth_id = OLVIS1A_DUCk
 version = 1
 provider = NSIDC_CUAT
-granule_regex = (?P<granuleid>.*_DUCK)\.JPG
+granule_regex = (?P<granuleid>.*_DUCk)\.JPG
 browse_regex = _reduced.JPG
 
 [Destination]
